refactor(datepicker): extract date conversion to helper utility
