+++ src/helpers/convertDateToString.ts
@@ -1,0 +1,23 @@
+export type DateFormat =
+  | "YYYY-MM-DD"
+  | "YYYY.MM.DD"
+  | "DD/MM/YYYY"
+  | "MM/DD/YYYY"
+  | "DD-MM-YYYY"
+  | "MM-DD-YYYY"
+  | "YYYY/MM/DD"
+  | "DD.MM.YYYY"
+  | "MM.DD.YYYY";
+
+export const convertToDateString = (date: Date, format: DateFormat): string => {
+  const day = String(date.getDate()).padStart(2, "0");
+  const month = String(date.getMonth() + 1).padStart(2, "0");
+  const year = date.getFullYear().toString();
+
+  const newDate = format
+    .replace("DD", day)
+    .replace("MM", month)
+    .replace("YYYY", year);
+
+  return newDate;
+};
